Add files via upload
Migration strategy of an e-commerce application to Azure: Managed Services on Microsoft Azure and Devsecops.

Devised a plan for migrating the on-premises e-commerce application to the Azure cloud, using the below phases. Also, listing the Azure services that are required in each phase.
Rehosting as-- Also referred to as "lift and shift," this stage entails migrating physical servers and VMs is to the cloud. By simply shifting your current server environment straight to Iaas. Analysis, Migration  and Optimization